fix: setting up the repository architecture
