+++ src/LandingPage.css
@@ -20,16 +20,6 @@
     background-color: rgba(0, 0, 0, 0.264);
     box-shadow: inset 10px 10px 130px 70px black;
   }
-  .header {
-    display: flex;
-    padding: 30px 60px;
-  }
-  .logo {
-    background-image: url(./img/Netflix_2015_logo.svg);
-    background-size: cover;
-    width: 170px;
-    height: 45px;
-  }
   
   .main {
     width: 100%;
@@ -162,44 +152,6 @@
   
   .kids-profiles__img {
     background-image: url(./img/kidsProfiles.png);
-  }
-  
-  .footer {
-    display: flex;
-    flex-direction: column;
-    background-color: black;
-    padding: 50px 0 30px 0;
-    align-items: center;
-  }
-  
-  .footer__title {
-    display: flex;
-    flex-direction: column;
-    margin: 20px 30px;
-  }
-  
-  a {
-    text-decoration: none;
-    color: #757575;
-  }
-  
-  a:hover {
-    text-decoration: underline;
-  }
-  .first-list li {
-    list-style-type: none;
-    margin: 16px 95px 0 0;
-    font-size: 13px;
-    cursor: pointer;
-  }
-  
-  .first-list {
-    display: grid;
-    grid-template-rows: 1fr 1fr 1fr 1fr;
-    grid-template-columns: 1fr 1fr 1fr 1fr;
-    text-align: start;
-    margin: 20px 0;
-    padding-inline-start: 0;
   }
   
   @media screen and (max-width: 756px) {
