+++ src/modules/Footer/Footer.css
@@ -1,0 +1,39 @@
+ 
+  .footer {
+    display: flex;
+    flex-direction: column;
+    background-color: black;
+    padding: 50px 0 30px 0;
+    align-items: center;
+  }
+  
+  .footer__title {
+    display: flex;
+    flex-direction: column;
+    margin: 20px 30px;
+  }
+  
+  a {
+    text-decoration: none;
+    color: #757575;
+  }
+  
+  a:hover {
+    text-decoration: underline;
+  }
+  .first-list li {
+    list-style-type: none;
+    margin: 16px 95px 0 0;
+    font-size: 13px;
+    cursor: pointer;
+  }
+  
+  .first-list {
+    display: grid;
+    grid-template-rows: 1fr 1fr 1fr 1fr;
+    grid-template-columns: 1fr 1fr 1fr 1fr;
+    text-align: start;
+    margin: 20px 0;
+    padding-inline-start: 0;
+  }
+
