+++ src/modules/Footer/Footer.jsx
@@ -1,0 +1,63 @@
+import "./Footer.css";
+
+
+function Footer() {
+  return (
+    <footer className="footer">
+      <div className="footer__title">
+        <a className="questions" href="#">
+          Questions? Contact us.
+        </a>
+        <ul className="first-list">
+          <li>
+            <a href="#">FAQ</a>
+          </li>
+          <li>
+            <a href="#">Help Center</a>
+          </li>
+          <li>
+            <a href="#">Account</a>
+          </li>
+          <li>
+            <a href="#">Media Center</a>
+          </li>
+          <li>
+            <a href="#">Investor Relations</a>
+          </li>
+          <li>
+            <a href="#">Jobs</a>
+          </li>
+          <li>
+            <a href="#">Ways to Watch</a>
+          </li>
+          <li>
+            <a href="#">Terms of Use</a>
+          </li>
+          <li>
+            <a href="#">Privacy</a>
+          </li>
+          <li>
+            <a href="#">Cookie Preferences</a>
+          </li>
+          <li>
+            <a href="#">Corporate Information</a>
+          </li>
+          <li>
+            <a href="#">Contact Us</a>
+          </li>
+          <li>
+            <a href="#">Speed Test</a>
+          </li>
+          <li>
+            <a href="#">Legal Notices</a>
+          </li>
+          <li>
+            <a href="#">Netflix Originals</a>
+          </li>
+        </ul>
+      </div>
+    </footer>
+  );
+}
+
+export default Footer;
