+++ src/modules/Modal.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+
+function Modal() {
+const a = document.querySelector(".header");
+
+        return ReactDOM.createPortal(
+        <div className="logo"></div>,
+a
+    )
+    }
+export default Modal;
